+++ test/theta_sketch_test.rb
@@ -1,8 +1,8 @@
 require_relative "test_helper"
 
 class ThetaSketchTest < Minitest::Test
-  def test_works
-    # sketch = DataSketches::ThetaSketch.new
-    # assert sketch.empty?
+  def test_update
+    sketch = DataSketches::UpdateThetaSketch.new
+    assert sketch.empty?
   end
 end